This is a java program that reads in a formated text file and spell checks the content
This spell checker reads the formated text file and determines if the
author is a high school, undergraduate or graduate student and grades
each text individually. The scoring for each student is different. The
user should expect a grade text file that contains misspelled words and
grades as an output.